feat: 이벤트 기반 아키텍처 도입
* refactor(payment): PgClient DIP 적용

* refactor: 패키지 위치 변경 및 분리하여 의존성 방향 정리

- PaymentV1Dto 는 외부 콜백 파라미터의 역할만 수행
- PgClient 요청에 대해서는 별도의 PgApiDto 를 생성하여 사용

* refactor: resilience4j 적용 위치 변경

- CardPaymentProcessor 에서 CardPaymentClient의 구현체인 PgClientAdapter 에서 재시도, 써킷브레이커 등을 적용하도록 변경
- 결제 요청과 조회 요청에 별도의 설정 적용
- FeignException을 별도의 커스텀 예외로 래핑

* refactor: 카드 결제 요청 리팩토링

- 외부 API 요청을 기존 트랜잭션으로부터 격리
  - 카드결제 전략은 이제 단순히 카드 결제 이벤트만 발행
  - CardPaymentService가 카드 결제 관련 로직을 담당

* refactor(like): 좋아요 처리에서 집계 로직 제거 및 이벤트 추가

* refactor(product): 재고차감 로직을 OrderFacade 에서 분리

* refactor(product): 결제 실패 시 재고 복원 로직 추가

- Order 에서 재고차감여부를 관리하고, 해당 값으로 재고를 복원할지 여부 결정

* refactor(product): 재고 복원 로직 이벤트 기반으로 변경

- 주문 서비스가 재고복원을 직접 하지 않도록 함

* refactor(coupon): 주문 실패 시 쿠폰 복원 로직 추가

+++ apps/commerce-api/src/main/java/com/loopers/domain/like/LikeService.java
@@ -4,8 +4,8 @@ import com.loopers.domain.product.ProductEntity;
 import com.loopers.domain.product.ProductRepository;
 import com.loopers.domain.user.UserEntity;
 import com.loopers.domain.user.UserRepository;
-import jakarta.persistence.EntityNotFoundException;
 import lombok.RequiredArgsConstructor;
+import org.springframework.context.ApplicationEventPublisher;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -19,33 +19,43 @@ public class LikeService {
     private final LikeRepository likeRepository;
     private final UserRepository userRepository;
     private final ProductRepository productRepository;
+    private final ApplicationEventPublisher eventPublisher;
 
     @Transactional
     public LikeEntity addLike(long userId, long productId) {
-        UserEntity user = userRepository.findById(userId)
-                .orElseThrow(() -> new EntityNotFoundException("사용자를 찾을 수 없습니다. userId: " + userId));
-        ProductEntity product = productRepository.findByIdWithPessimisticLock(productId)
-                .orElseThrow(() -> new EntityNotFoundException("상품을 찾을 수 없습니다. productId: " + productId));
+        return likeRepository.findByUserIdAndProductId(userId, productId)
+                // 값이 있는 경우 (이미 좋아요를 누른 적이 있음)
+                .map(like -> {
+                    // 삭제된 상태일 때만 복원하고 이벤트를 발행
+                    if (like.isDeleted()) {
+                        like.restore();
+                        likeRepository.save(like);
+                        eventPublisher.publishEvent(new LikeChangedEvent(productId, userId, true));
+                    }
+                    return like;
+                })
+                // 값이 없는 경우 (처음으로 좋아요를 누름)
+                .orElseGet(() -> {
+                    UserEntity userProxy = userRepository.getReferenceById(userId);
+                    ProductEntity productProxy = productRepository.getReferenceById(productId);
+                    LikeEntity newLike = LikeEntity.create(userProxy, productProxy);
 
-        LikeEntity newLike = likeRepository.saveOrFind(LikeEntity.create(user, product));
-
-        product.increaseLikeCount();
-        productRepository.save(product);
-
-        return newLike;
+                    likeRepository.save(newLike);
+                    eventPublisher.publishEvent(new LikeChangedEvent(productId, userId, true));
+                    return newLike;
+                });
     }
 
     @Transactional
     public void removeLike(long userId, long productId) {
         likeRepository.findByUserIdAndProductId(userId, productId)
                 .ifPresent(like -> {
+                    if (like.isDeleted()) {
+                        return;
+                    }
                     like.delete();
                     likeRepository.save(like);
-
-                    ProductEntity product = productRepository.findByIdWithPessimisticLock(productId)
-                            .orElseThrow(() -> new EntityNotFoundException("상품을 찾을 수 없습니다. productId: " + productId));
-                    product.decreaseLikeCount();
-                    productRepository.save(product);
+                    eventPublisher.publishEvent(new LikeChangedEvent(productId, userId, false));
                 });
     }
